Lesson 1: Notes
If the value inside the object is a function, we can use concise method notation. This is part of ES6 and only available when a function is defined inside of an object

When we use arrow functions for defining action creators and the only statement after the function block is return, we can use the object that is returning as the body of the arrow function. This way instead of using a block we use the object expression.

To help the parser understand we use parenthesis