Lesson 02: Notes
When Redux initializes, by default the state argument is `undefined` and is replaced by the default value the root reducer(s) defines.

However, if we have existing data and pass it as a second argunment of `createStore()`, we essentially are giving the root reducer a state argument that's different from `undefined`. This causes the default argument sintax in the root reducer to fail and instead, is replaced by the state argument that was passed to the store.

This causes an override of the reducer's value.

In this case, since we have the todos value inside the persistedState object, when the store is created this value will override the default value defined by the reducer.

It's good practice to use this second argument to get persistend data from store, but not for specify initial state tree of our app. Using reducers for initial state, makes it easy for changes and testing